kweetservice fully functional, gateway routing fix, customer insertion

+++ Kwetter/Services/FeedService/API/Eventing/EventConsumer/CustomerFollowed/CustomerFollowedEvent.cs
@@ -4,8 +4,8 @@ namespace FeedService.API.Eventing.EventConsumer.CustomerFollowed
 {
     public class CustomerFollowedEvent : Event
     {
-        public long FollowerId { get; set; }
-        public long FollowingId { get; set; }
+        public int FollowerId { get; set; }
+        public int FollowingId { get; set; }
         public DateTime FollowedDateTime { get; set; }
     }
 }
