+++ Kwetter/Services/FeedService/API/Eventing/EventConsumer/CustomerUnfollowed/CustomerUnfollowedEvent.cs
@@ -4,7 +4,7 @@ namespace FeedService.API.Eventing.EventConsumer.CustomerUnfollowed
 {
     public class CustomerUnfollowedEvent : Event
     {
-        public long FollowerId { get; set; }
-        public long FollowingId { get; set; }
+        public int FollowerId { get; set; }
+        public int FollowingId { get; set; }
     }
 }
