+++ Kwetter/Services/FeedService/API/Eventing/EventConsumer/KweetCreated/KweetCreatedEvent.cs
@@ -4,8 +4,8 @@ namespace FeedService.API.Eventing.EventConsumer.KweetCreated
 {
     public class KweetCreatedEvent : Event
     {
-        public long KweetId { get; set; }
-        public long CustomerId { get; set; }
+        public int KweetId { get; set; }
+        public int CustomerId { get; set; }
         public string Text { get; set; }
         public DateTime KweetCreatedDate { get; set; }
     }
