+++ Kwetter/Services/FeedService/API/Eventing/EventConsumer/KweetLiked/KweetLikedEvent.cs
@@ -4,8 +4,8 @@ namespace FeedService.API.Eventing.EventConsumer.KweetLiked
 {
     public class KweetLikedEvent : Event
     {
-        public long KweetId { get; set; }
-        public long CustomerId { get; set; }
+        public int KweetId { get; set; }
+        public int CustomerId { get; set; }
         public string Text { get; set; }
         public DateTime KweetCreatedDate { get; set; }
     }
