+++ Kwetter/Services/KweetService/API/Eventing/EventPublisher/KweetUnliked/KweetUnlikedEvent.cs
@@ -4,9 +4,9 @@ namespace KweetService.API.Eventing.EventPublisher.KweetUnliked
 {
     public class KweetUnlikedEvent : IRequest
     {
-        public long LikeId { get; set; }
-        public long KweetId { get; set; }
-        public long CustomerId { get; set; }
+        public int LikeId { get; set; }
+        public int KweetId { get; set; }
+        public int CustomerId { get; set; }
         public DateTime LikedDateTime { get; set; }
     }
 }
