+++ Kwetter/Services/KweetService/API/Eventing/EventPublisher/Kweetliked/KweetLikedEvent.cs
@@ -4,9 +4,9 @@ namespace KweetService.API.Eventing.EventPublisher.KweetLiked
 {
     public class KweetLikedEvent : IRequest
     {
-        public long LikeId { get; set; }
-        public long KweetId { get; set; }
-        public long CustomerId { get; set; }
+        public int LikeId { get; set; }
+        public int KweetId { get; set; }
+        public int CustomerId { get; set; }
         public DateTime LikedDateTime { get; set; }
     }
 }
